fix: v0.39.1 - Fix DuplicateStats type mismatch causing template error
Changed CrossDiskGroups and SameDiskGroups from int to int64 to match
the template 'add' function signature, which expects int64 parameters.

This fixes the template error:
'wrong type for value, expected int64; got int'

Version: 0.39.0 → 0.39.1

+++ internal/database/duplicates.go
@@ -323,8 +323,8 @@ func (db *DB) GetDuplicateStats() (*DuplicateStats, error) {
 
 // DuplicateStats contains summary statistics about duplicates
 type DuplicateStats struct {
-	CrossDiskGroups           int
-	SameDiskGroups            int
+	CrossDiskGroups           int64
+	SameDiskGroups            int64
 	CrossDiskPotentialSavings int64
 	SameDiskPotentialSavings  int64
 	TotalPotentialSavings     int64
